# pipeline: use Message value to reduce GC pressure