--- a/pkg/pipeline/message.go
+++ b/pkg/pipeline/message.go
@@ -41,39 +41,35 @@ type Message struct {
 }
 
 // PolymorphicEventMessage creates the message of PolymorphicEvent
-func PolymorphicEventMessage(event *model.PolymorphicEvent) *Message {
-	return &Message{
+func PolymorphicEventMessage(event *model.PolymorphicEvent) Message {
+	return Message{
 		Tp:               MessageTypePolymorphicEvent,
 		PolymorphicEvent: event,
 	}
 }
 
 // CommandMessage creates the message of Command
-func CommandMessage(command *Command) *Message {
-	return &Message{
+func CommandMessage(command *Command) Message {
+	return Message{
 		Tp:      MessageTypeCommand,
 		Command: command,
 	}
 }
 
 // BarrierMessage creates the message of Command
-func BarrierMessage(barrierTs model.Ts) *Message {
-	return &Message{
+func BarrierMessage(barrierTs model.Ts) Message {
+	return Message{
 		Tp:        MessageTypeBarrier,
 		BarrierTs: barrierTs,
 	}
 }
 
-// TickMessage is called frequently,
-// to ease GC pressure we return a global variable.
-var tickMsg *Message = &Message{
-	Tp: MessageTypeTick,
-}
-
 // TickMessage creates the message of Tick
 // Note: the returned message is READ-ONLY.
-func TickMessage() *Message {
-	return tickMsg
+func TickMessage() Message {
+	return Message{
+		Tp: MessageTypeTick,
+	}
 }
 
 // CommandType is the type of Command