--- a/pkg/pipeline/test.go
+++ b/pkg/pipeline/test.go
@@ -19,18 +19,18 @@ import "github.com/pingcap/ticdc/pkg/context"
 // This function is only for testing.
 // Only `Node.Receive` will be called, other functions in `Node` will never be called.
 // When the `Receive` function of the `Node` returns an error, this function will return the message that caused the error and the error
-func SendMessageToNode4Test(ctx context.Context, node Node, msgs []*Message, outputCh chan *Message) (*Message, error) {
-	nodeCtx := newNodeContext(ctx, nil, outputCh)
+func SendMessageToNode4Test(ctx context.Context, node Node, msgs []Message, outputCh chan Message) (Message, error) {
+	nodeCtx := newNodeContext(ctx, Message{}, outputCh)
 	for _, msg := range msgs {
 		err := node.Receive(withMessage(nodeCtx, msg))
 		if err != nil {
 			return msg, err
 		}
 	}
-	return nil, nil
+	return Message{}, nil
 }
 
 // MockNodeContext4Test creates a node context with a message and a output channel for tests.
-func MockNodeContext4Test(ctx context.Context, msg *Message, outputCh chan *Message) NodeContext {
+func MockNodeContext4Test(ctx context.Context, msg Message, outputCh chan Message) NodeContext {
 	return newNodeContext(ctx, msg, outputCh)
 }