### What problem does this PR solve? <!--add issue link with summary if exists-->

Use Message value to reduce GC pressure

```
# master
goos: linux
goarch: amd64
pkg: github.com/pingcap/ticdc/pkg/pipeline
cpu: Intel(R) Xeon(R) CPU E5-2630 v4 @ 2.20GHz
BenchmarkPipeline/BenchmarkPipeline/1_node(s)-40         	 1309424	       819.7 ns/op	      56 B/op	       2 allocs/op
BenchmarkPipeline/BenchmarkPipeline/2_node(s)-40         	  953646	      1214 ns/op	      80 B/op	       3 allocs/op
BenchmarkPipeline/BenchmarkPipeline/3_node(s)-40         	  739050	      1639 ns/op	     104 B/op	       4 allocs/op
BenchmarkPipeline/BenchmarkPipeline/4_node(s)-40         	  631210	      1953 ns/op	     128 B/op	       5 allocs/op
BenchmarkPipeline/BenchmarkPipeline/5_node(s)-40         	  535172	      2387 ns/op	     152 B/op	       6 allocs/op
BenchmarkPipeline/BenchmarkPipeline/6_node(s)-40         	  429428	      2819 ns/op	     176 B/op	       7 allocs/op
BenchmarkPipeline/BenchmarkPipeline/7_node(s)-40         	  404947	      3508 ns/op	     200 B/op	       8 allocs/op
BenchmarkPipeline/BenchmarkPipeline/8_node(s)-40         	  326889	      3714 ns/op	     224 B/op	       9 allocs/op
PASS
ok  	github.com/pingcap/ticdc/pkg/pipeline	11.001s

# this patch
goos: linux
goarch: amd64
pkg: github.com/pingcap/ticdc/pkg/pipeline
cpu: Intel(R) Xeon(R) CPU E5-2630 v4 @ 2.20GHz
BenchmarkPipeline/BenchmarkPipeline/1_node(s)-40         	 1643746	       732.4 ns/op	      48 B/op	       1 allocs/op
BenchmarkPipeline/BenchmarkPipeline/2_node(s)-40         	  972543	      1195 ns/op	      96 B/op	       2 allocs/op
BenchmarkPipeline/BenchmarkPipeline/3_node(s)-40         	  777621	      1634 ns/op	     144 B/op	       3 allocs/op
BenchmarkPipeline/BenchmarkPipeline/4_node(s)-40         	  648489	      1982 ns/op	     192 B/op	       4 allocs/op
BenchmarkPipeline/BenchmarkPipeline/5_node(s)-40         	  507319	      2488 ns/op	     240 B/op	       5 allocs/op
BenchmarkPipeline/BenchmarkPipeline/6_node(s)-40         	  395040	      2822 ns/op	     288 B/op	       6 allocs/op
BenchmarkPipeline/BenchmarkPipeline/7_node(s)-40         	  376586	      3316 ns/op	     336 B/op	       7 allocs/op
BenchmarkPipeline/BenchmarkPipeline/8_node(s)-40         	  318111	      3858 ns/op	     384 B/op	       8 allocs/op
PASS
ok  	github.com/pingcap/ticdc/pkg/pipeline	10.624s
```

Left: master
Right: this patch
![image](https://user-images.githubusercontent.com/2150711/127456653-f6fcd47d-0a56-4b38-b77a-48b2936e6086.png)
![image](https://user-images.githubusercontent.com/2150711/127456725-5dfd6a44-2b7d-4fd3-96b1-ab5956c7cc1e.png)
![image](https://user-images.githubusercontent.com/2150711/127456755-079b066e-1930-491c-b1d9-9ebc7ce0d8f8.png)


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

Related changes

 - Need to cherry-pick to the release branch

### Release note <!-- bugfixes or new feature need a release note -->

```release-note
None
```